refactor: modularize plugin runtime and test hooks
(cherry picked from commit 0fd77c9856a22e01f010de3eef7439cf82572e80)
