+++ src/plugins/runtime/runtime-config.ts
@@ -1,0 +1,9 @@
+import { loadConfig, writeConfigFile } from "../../config/config.js";
+import type { PluginRuntime } from "./types.js";
+
+export function createRuntimeConfig(): PluginRuntime["config"] {
+  return {
+    loadConfig,
+    writeConfigFile,
+  };
+}
