+++ src/plugins/runtime/runtime-logging.ts
@@ -1,0 +1,21 @@
+import { shouldLogVerbose } from "../../globals.js";
+import { getChildLogger } from "../../logging.js";
+import { normalizeLogLevel } from "../../logging/levels.js";
+import type { PluginRuntime } from "./types.js";
+
+export function createRuntimeLogging(): PluginRuntime["logging"] {
+  return {
+    shouldLogVerbose,
+    getChildLogger: (bindings, opts) => {
+      const logger = getChildLogger(bindings, {
+        level: opts?.level ? normalizeLogLevel(opts.level) : undefined,
+      });
+      return {
+        debug: (message) => logger.debug?.(message),
+        info: (message) => logger.info(message),
+        warn: (message) => logger.warn(message),
+        error: (message) => logger.error(message),
+      };
+    },
+  };
+}
