+++ src/plugins/runtime/runtime-media.ts
@@ -1,0 +1,17 @@
+import { isVoiceCompatibleAudio } from "../../media/audio.js";
+import { mediaKindFromMime } from "../../media/constants.js";
+import { getImageMetadata, resizeToJpeg } from "../../media/image-ops.js";
+import { detectMime } from "../../media/mime.js";
+import { loadWebMedia } from "../../web/media.js";
+import type { PluginRuntime } from "./types.js";
+
+export function createRuntimeMedia(): PluginRuntime["media"] {
+  return {
+    loadWebMedia,
+    detectMime,
+    mediaKindFromMime,
+    isVoiceCompatibleAudio,
+    getImageMetadata,
+    resizeToJpeg,
+  };
+}
