+++ src/plugins/runtime/runtime-system.ts
@@ -1,0 +1,14 @@
+import { requestHeartbeatNow } from "../../infra/heartbeat-wake.js";
+import { enqueueSystemEvent } from "../../infra/system-events.js";
+import { runCommandWithTimeout } from "../../process/exec.js";
+import { formatNativeDependencyHint } from "./native-deps.js";
+import type { PluginRuntime } from "./types.js";
+
+export function createRuntimeSystem(): PluginRuntime["system"] {
+  return {
+    enqueueSystemEvent,
+    requestHeartbeatNow,
+    runCommandWithTimeout,
+    formatNativeDependencyHint,
+  };
+}
