+++ src/plugins/runtime/runtime-tools.ts
@@ -1,0 +1,11 @@
+import { createMemoryGetTool, createMemorySearchTool } from "../../agents/tools/memory-tool.js";
+import { registerMemoryCli } from "../../cli/memory-cli.js";
+import type { PluginRuntime } from "./types.js";
+
+export function createRuntimeTools(): PluginRuntime["tools"] {
+  return {
+    createMemoryGetTool,
+    createMemorySearchTool,
+    registerMemoryCli,
+  };
+}
